decrating: absorb shipper-environment into shipper::runtime::environment
Stacks on PR #53 (env+storage+git dual-impl cleanup). Move the 2202-LOC
shipper-environment crate into crates/shipper/src/runtime/environment/
as crate-private modules (mod.rs + fingerprint.rs + ci.rs, ~2190 LOC
total). Preserves the deduped in-tree wrapper logic (normalize_version,
EnvironmentInfo::collect delegation with graceful fallback) from PR #53.

Split seams:
- mod.rs         — CiEnvironment, detect_environment, is_ci,
                   collect_environment_fingerprint (the absorbed shim),
                   CiEnvironment snapshot tests.
- fingerprint.rs — EnvironmentInfo, get_rust_version, get_cargo_version,
                   get_environment_fingerprint, normalize_tool_version,
                   collect_env_vars, tests + proptests.
- ci.rs          — get_ci_branch, get_ci_commit_sha, is_pull_request,
                   provider-specific tests.

All 13 snapshots preserved (renamed to `shipper__runtime__environment__
snapshot_tests__*.snap`). All 130 absorbed tests pass in-tree.

Surgical unblock (precedent: PR #54 policy→engine-parallel inline):
shipper-state depended on shipper-environment via one call site in
migrate_v1_to_v2. Because shipper-state is an (optional) dep of
shipper, it cannot depend on shipper itself — a dep cycle. Inlined an
~20-LOC collect_environment_fingerprint into shipper-state (marked
`#[doc(hidden)]`, with a comment pointing back at
crate::runtime::environment for the post-shipper-state-absorption
replacement), and dropped shipper-environment from shipper-state's
Cargo.toml. Will be replaced with a direct call to
crate::runtime::environment::collect_environment_fingerprint after
shipper-state is itself absorbed into shipper::state.

Module is `pub(crate) mod runtime; mod environment;` — the former
public `shipper::environment` API had no external consumers in-tree
and is now crate-private per the runtime-layer discipline.
`#[allow(dead_code)]` is applied at the runtime::environment level
because several absorbed items (CI branch/SHA/PR helpers, the short
pipe-fingerprint form, EnvironmentInfo::fingerprint) currently have
no in-crate callers but retain full test coverage for future wiring.

Also updates shipper's `insta` dev-dep to enable the `yaml` feature
(required by the moved snapshot_tests module).

Per docs/decrating-plan.md section 6 Phase 2.
