--- crates/shipper-environment/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "shipper-environment"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-rust-version.workspace = true
-description = "Environment fingerprinting for shipper"
-repository = "https://github.com/EffortlessMetrics/shipper"
-documentation = "https://docs.rs/shipper-environment"
-keywords = ["environment", "fingerprint", "ci"]
-categories = ["development-tools"]
-
-[dependencies]
-serde.workspace = true
-anyhow = "1.0"
-chrono = { version = "0.4.44", features = ["serde"] }
-shipper-types = { path = "../shipper-types" }
-
-[dev-dependencies]
-tempfile = "3"
-serde_json = "1.0"
-serial_test = "3"
-temp-env = "0.3"
-proptest.workspace = true
-insta.workspace = true
-
-[lints]
-workspace = true
