+++ crates/shipper-state/Cargo.toml
@@ -16,7 +16,6 @@ serde_json = "1.0"
 anyhow = "1.0"
 chrono = { version = "0.4.44", features = ["serde"] }
 shipper-types = { path = "../shipper-types" }
-shipper-environment = { path = "../shipper-environment" }
 shipper-encrypt = { path = "../shipper-encrypt" }
 shipper-schema = { path = "../shipper-schema" }
