+++ crates/shipper/src/runtime/mod.rs
@@ -5,3 +5,11 @@
 
 pub(crate) mod policy;
 pub mod execution;
+
+// Some absorbed `environment` items (CI branch/SHA/PR helpers, pipe-fingerprint
+// form, `normalize_tool_version`, `EnvironmentInfo::fingerprint`) currently
+// have no in-crate callers but have full test coverage. They were public API
+// of the former `shipper-environment` microcrate and are kept available for
+// future wiring (e.g., webhook/event metadata) rather than dropped.
+#[allow(dead_code)]
+pub(crate) mod environment;
